Django_backend/requirements.txt
Django>=5.2.8
mysqlclient>=2.2.0
PyJWT>=2.8.0

